Quicjs optimizes similar to engine; objects lerp from their prototype

+++ source/engine/resources.h
@@ -16,7 +16,7 @@ char *make_path(const char *file);
 
 char *strdup(const char *s);
 
-unsigned char *slurp_file(const char *filename, long *size);
+unsigned char *slurp_file(const char *filename, size_t *size);
 char *slurp_text(const char *filename);
 int slurp_write(const char *txt, const char *filename);
